2017/02/09 fhr study 2/3

+++ MusicStore/MusicStore/Models/Artist.cs
@@ -5,14 +5,9 @@ using System.Web;
 
 namespace MusicStore.Models
 {
-    /// <summary>
-    /// 艺术家
-    /// 2017/1/20 fhr
-    /// </summary>
     public class Artist
     {
-        public Int64 ArtistId { get; set; }
-
+        public int ArtistId { get; set; }
         public string Name { get; set; }
     }
 }
